+++ lib/wingify/models/user/context_model.rb
@@ -17,15 +17,16 @@ require_relative '../../utils/uuid_util'
 require_relative '../../services/settings_service'
 
 class ContextModel
-  attr_accessor :id, :user_agent, :ip_address, :custom_variables, :variation_targeting_variables, :post_segmentation_variables, :vwo, :session_id, :uuid, :bucketing_seed
+  attr_accessor :id, :user_agent, :ip_address, :custom_variables, :variation_targeting_variables, :post_segmentation_variables, :platform_variables, :vwo, :session_id, :uuid, :bucketing_seed
 
-  def initialize(id = nil, user_agent = nil, ip_address = nil, custom_variables = {}, variation_targeting_variables = {}, post_segmentation_variables = {}, vwo = nil, session_id = nil, uuid = nil, bucketing_seed = nil)
+  def initialize(id = nil, user_agent = nil, ip_address = nil, custom_variables = {}, variation_targeting_variables = {}, post_segmentation_variables = {}, platform_variables = {}, vwo = nil, session_id = nil, uuid = nil, bucketing_seed = nil)
     @id = id
     @user_agent = user_agent
     @ip_address = ip_address
     @custom_variables = custom_variables || {}
     @variation_targeting_variables = variation_targeting_variables || {}
     @post_segmentation_variables = post_segmentation_variables || {}
+    @platform_variables = platform_variables || {}
     @vwo = vwo
     @session_id = session_id
     @uuid = uuid
@@ -52,6 +53,7 @@ class ContextModel
     @custom_variables = context[:customVariables] if context.key?(:customVariables)
     @variation_targeting_variables = context[:variationTargetingVariables] if context.key?(:variationTargetingVariables)
     @post_segmentation_variables = context[:postSegmentationVariables] if context.key?(:postSegmentationVariables)
+    @platform_variables = context[:platformVariables] if context.key?(:platformVariables)
     @vwo = ContextVWOModel.new.model_from_dictionary(context[:_vwo]) if context.key?(:_vwo)
     @bucketing_seed = context[:bucketingSeed].to_s if context.key?(:bucketingSeed)
 
@@ -118,6 +120,14 @@ class ContextModel
 
   def set_post_segmentation_variables(post_segmentation_variables)
     @post_segmentation_variables = post_segmentation_variables
+  end
+
+  def get_platform_variables
+    @platform_variables
+  end
+
+  def set_platform_variables(platform_variables)
+    @platform_variables = platform_variables
   end
 
   def get_vwo