+++ lib/wingify/packages/segmentation_evaluator/enums/segment_operator_value_enum.rb
@@ -30,4 +30,5 @@ module SegmentOperatorValueEnum
     IP = 'ip_address'
     BROWSER_VERSION = 'browser_version'
     OS_VERSION = 'os_version'
+    WEB_CAMPAIGN_VARIATION = 'campaignVariation'
   end